Move ElfBackend type into separate module
This change moves the ElfBackend type into a separate module, where it
will reside for the time being. It has no business being located in a
module called 'cache'.
