+++ src/elf/mod.rs
@@ -1,3 +1,4 @@
+mod backend;
 #[cfg_attr(not(feature = "dwarf"), allow(unused_variables))]
 mod cache;
 mod parser;
@@ -6,7 +7,7 @@ mod resolver;
 #[allow(dead_code, non_camel_case_types)]
 pub(crate) mod types;
 
-pub(crate) use cache::ElfBackend;
+pub(crate) use backend::ElfBackend;
 pub(crate) use cache::ElfCache;
 pub(crate) use parser::ElfParser;
 pub(crate) use resolver::ElfResolver;
